Moves to separate module and renames, also simplifies overwriting of parameters
