+++ pyfilter/inference/kernels/online.py
@@ -1,25 +1,8 @@
 from .base import BaseKernel
 from .kde import KernelDensityEstimate, NonShrinkingKernel
-from ..utils import stacker
 from ...utils import unflattify, get_ess
 import torch
 from ...filters import BaseFilter, BaseState
-
-
-def _cont_jitter(parameters, stacked, jittered, ess):
-    for p, msk, ps in zip(parameters, stacked.mask, stacked.prev_shape):
-        p.t_values = unflattify(jittered[:, msk], ps)
-
-    return
-
-
-def _disc_jitter(parameters, stacked, jittered, ess):
-    to_jitter = torch.empty(jittered.shape[0], device=jittered.device).bernoulli_(1 / ess ** 0.5)
-
-    for p, msk, ps in zip(parameters, stacked.mask, stacked.prev_shape):
-        p.t_values = (1 - to_jitter) * p.t_values + to_jitter * unflattify(jittered[:, msk], ps)
-
-    return
 
 
 class OnlineKernel(BaseKernel):
@@ -32,7 +15,7 @@ class OnlineKernel(BaseKernel):
         super().__init__(**kwargs)
 
         self._kde = kde or NonShrinkingKernel()
-        self._mutater = _cont_jitter if not discrete else _disc_jitter
+        self._disc = discrete
 
     def _resample(self, filter_: BaseFilter, state: BaseState, weights: torch.Tensor):
         """
@@ -48,13 +31,20 @@ class OnlineKernel(BaseKernel):
         return inds
 
     def _update(self, parameters, filter_, state, weights):
-        # ===== Perform shrinkage ===== #
-        stacked = stacker(parameters, lambda u: u.t_values)
+        stacked = filter_.ssm.parameters_as_matrix(transformed=True)
         kde = self._kde.fit(stacked.concated, weights)
 
         inds = self._resample(filter_, state, weights)
         jittered = kde.sample(inds=inds)
 
-        self._mutater(parameters, stacked, jittered, get_ess(weights, normalized=True))
+        if self._disc:
+            to_jitter = torch.empty(
+                jittered.shape[0], device=jittered.device
+            ).bernoulli_(1 / get_ess(weights, True) ** 0.5)
+
+            jittered = (1 - to_jitter) * stacked.concated[inds] + to_jitter * jittered
+
+        new_params = tuple(unflattify(jittered[:, msk], ps) for msk, ps in zip(stacked.mask, stacked.prev_shape))
+        filter_.ssm.update_parameters(new_params)
 
         return self
