+++ pyfilter/timeseries/model.py
@@ -1,11 +1,12 @@
-from ..utils import flatten
+from ..utils import flatten, stacker
 import torch
-from .base import StochasticProcess, StochasticProcessBase
+from .base import Base
+from .process import StochasticProcess
 from typing import Tuple
 from .parameter import Parameter
 
 
-class StateSpaceModel(StochasticProcessBase):
+class StateSpaceModel(Base):
     def __init__(self, hidden: StochasticProcess, observable: StochasticProcess):
         """
         Combines a hidden and observable processes to constitute a state-space model.
@@ -45,6 +46,17 @@ class StateSpaceModel(StochasticProcessBase):
 
     def viewify_params(self, shape, in_place=True) -> Tuple[Tuple[Parameter, ...], ...]:
         return tuple(ssm.viewify_params(shape, in_place=in_place) for ssm in [self.hidden, self.observable])
+
+    def update_parameters(self, params, transformed=True):
+        num_params = len(self.hidden.theta_dists)
+
+        for m, ps in [(self.hidden, params[:num_params]), (self.observable, params[num_params:])]:
+            m.update_parameters(ps)
+
+        return self
+
+    def parameters_as_matrix(self, transformed=True):
+        return stacker(self.theta_dists, lambda u: u.t_values if transformed else u.values)
 
     def h_weight(self, y: torch.Tensor, x: torch.Tensor):
         """
